@@ -4,7 +4,6 @@ import com.example.mitch.pmanager.interfaces.Writable;
 
 import java.io.Serializable;
 import java.util.ArrayList;
-import java.util.HashMap;
 
 /**
  * PasswordBank class for new file storage schema
@@ -14,51 +13,61 @@ public class PasswordBank implements Serializable, Writable {
     /**
      * Hashmap of data to store
      */
-    private final HashMap<String, ArrayList<UserEntry>> bank;
+    private final ArrayList<DomainEntry> bank;
 
-    /**
-     * Default constructor
-     * Creates a new hashmap
-     */
     public PasswordBank() {
-        bank = new HashMap<>();
+        bank = new ArrayList<>();
     }
 
     /**
-     * Hashmap constructor
+     * ArrayList constructor
+     *
      * @param bank Bank to use
      */
-    public PasswordBank(HashMap<String, ArrayList<UserEntry>> bank) {
+    public PasswordBank(ArrayList<DomainEntry> bank) {
         this.bank = bank;
     }
 
     /**
-     * Gets the arraylist of entries for the given domain
-     * @param domain Domain to retrieve entries for
-     * @return ArrayList of entries
+     * Gets the DomainEntry for a given domain
+     *
+     * @param domain Domain to retrieve DomainEntry for
+     * @return DomainEntry of given domain
      */
-    public ArrayList<UserEntry> getEntries(String domain) {
-        return bank.get(domain);
+    public DomainEntry getEntry(String domain) {
+        for (DomainEntry domainEntry : bank) {
+            if (domainEntry.getDomain().equals(domain)) return domainEntry;
+        }
+        return null;
     }
 
     /**
      * Creates an empty domain with a new ArrayList
+     *
      * @param domain Domain to create
-     * @return ArrayList for later use
+     * @return DomainEntry for later use
      */
-    public ArrayList<UserEntry> createDomain(String domain) {
-        ArrayList<UserEntry> entries = new ArrayList<>();
-        bank.put(domain, entries);
-        return entries;
+    public DomainEntry createDomain(String domain) {
+        for (DomainEntry domainEntry : bank) {
+            if (domainEntry.getDomain().equals(domain)) return domainEntry;
+        }
+        DomainEntry entry = new DomainEntry(new ArrayList<>(), domain);
+        bank.add(entry);
+        return entry;
     }
 
     /**
      * Sets the given domain to the given entries list
-     * @param domain Domain to set
+     *
+     * @param domain  Domain to set
      * @param entries ArrayList for the domain
      */
     public void addEntries(String domain, ArrayList<UserEntry> entries) {
-        bank.put(domain, entries);
+        createDomain(domain).getEntries().addAll(entries);
+    }
+
+    public ArrayList<DomainEntry> getEntries() {
+        return bank;
     }
 
     private static final long serialVersionUID = 1L;
